rename CategoryPage to CategoryUiDriver
