+++ e2e/src/domains/category/infrastructure/pages/category-updating.page.ts
@@ -33,8 +33,6 @@ export class CategoryUpdatingPage extends BasePage {
     this.logger.info(
       `Starting updateCategoryParent with categoryName: '${categoryName}', newParentName: '${newParentName}'`
     );
-    // Ensure on Category Management page
-    await this.assertOnCategoryPage();
 
     // Find the row/container that contains the category name
     const categoryRows = this.page.locator(
@@ -106,31 +104,5 @@ export class CategoryUpdatingPage extends BasePage {
     this.logger.info(
       `Ending updateCategoryParent with categoryName: '${categoryName}', newParentName: '${newParentName}'`
     );
-  }
-
-  async assertOnCategoryPage(): Promise<void> {
-    const url = this.page.url();
-    this.logger.info(`url: ${url}`);
-
-    if (!(await url).includes("/categories")) {
-      throw new Error("User is not on the Category Management page");
-    }
-    // Wait for network to be idle
-    await this.page.waitForLoadState("networkidle");
-    // Wait for loading overlay to disappear if it exists
-    try {
-      await this.page.waitForSelector("[data-test=loading-overlay]", {
-        state: "detached",
-        timeout: 5000,
-      });
-    } catch (e) {
-      // If overlay never appears, that's fine
-    }
-    const isVisible = await this.page.isVisible(
-      "[data-testid=add-category-button]"
-    );
-    if (!isVisible) {
-      throw new Error("Category Management page did not load correctly");
-    }
   }
 }
